[Refact] #12 - RemoteData와 Repository 수정
- 기존 SignUp RemoteData와 Repository Auth로 통합 중
