+++ app/src/main/java/org/shop/yogizogi_android/data/model/remote/datasource/ProcessCall.kt
@@ -1,5 +1,6 @@
 package org.shop.yogizogi_android.data.model.remote.datasource
 
+import android.util.Log
 import com.google.gson.Gson
 import org.shop.yogizogi_android.data.Resource
 import org.shop.yogizogi_android.data.model.remote.response.CommonFailRes
@@ -16,8 +17,9 @@ suspend fun <T> processCall(call: suspend () -> Response<CommonSuccessRes<T>>): 
             Resource.Error("Server error", null)
         }
     } else {
+        Log.d("processCall - FailRes", response.toString())
         val commonFailRes = response.errorBody()?.let {
-            Gson().fromJson(it.string(), CommonFailRes::class.java)
+            Gson().fromJson(it.toString(), CommonFailRes::class.java)
         }
         val errorMessage = commonFailRes?.message ?: "Server error"
         Resource.Error(errorMessage, null)
