+++ app/src/main/java/org/shop/yogizogi_android/di/DataModule.kt
@@ -6,8 +6,6 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import org.shop.yogizogi_android.repository.AuthRepository
 import org.shop.yogizogi_android.repository.AuthRepositoryImpl
-import org.shop.yogizogi_android.repository.SignUpRepository
-import org.shop.yogizogi_android.repository.SignUpRepositoryImpl
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -15,7 +13,4 @@ interface DataModule {
 
     @Binds
     fun provideAuthRepository(authRepositoryImpl: AuthRepositoryImpl): AuthRepository
-
-    @Binds
-    fun provideSignUpRepository(signUpRepositoryImpl: SignUpRepositoryImpl): SignUpRepository
 }
