+++ app/src/main/java/org/shop/yogizogi_android/repository/AuthRepository.kt
@@ -2,7 +2,9 @@ package org.shop.yogizogi_android.repository
 
 import kotlinx.coroutines.flow.Flow
 import org.shop.yogizogi_android.data.Resource
+import org.shop.yogizogi_android.data.model.remote.request.SignUpReqDTO
 import org.shop.yogizogi_android.data.model.remote.response.LogInResDTO
+import org.shop.yogizogi_android.data.model.remote.response.SignUpResDTO
 import org.shop.yogizogi_android.data.model.remote.response.VerifyCodeCheckResDTO
 import org.shop.yogizogi_android.data.model.remote.response.VerifyCodeSendResDTO
 
@@ -13,4 +15,5 @@ interface AuthRepository {
         phoneNumber: String,
         code: String
     ): Flow<Resource<VerifyCodeCheckResDTO>>
+    suspend fun postSignUp(body: SignUpReqDTO): Flow<Resource<SignUpResDTO>>
 }
